+++ app.js
@@ -1,19 +1,10 @@
-const mymodule = require('./index.js')
+const { readMd } = require ('./utils/mdFilter.js')
 
 console.log('o---o---o---o---o---o---o---o---o')
 console.log('  Bienvenido a Md-Lnks   0-|--<  ')
 console.log('o---o---o---o---o---o---o---o---o')
 
-/* mymodule("./Directorio", function(err, files){
-    if (err){
-        console.log(err)
-    } else {
-        files.forEach(function(fileName){
-            console.log(fileName)
-        })
-    }
-})  */
 
-mymodule("./Directorio")
-    .then( files => console.log(files))
-    .catch( err => console.log(err))
+
+
+console.log(readMd('./Directorio'))
